// src/dtos/PeopleResultsEs.dto.ts
class PeopleResultsEs  {
  
  nombre: string;
  altura: string;
  masa: string;
  color_cabello: string;
  color_piel: string;
  color_ojos: string;
  anio_nacimiento: string;
  genero: string;
  mundo_natal: string;
  peliculas?: any[];
  especies?: any[];
  vehiculos?: any[];
  naves_estelares?: any[];
  url: string;
  creado: string;
  editado: string;
}

export default PeopleResultsEs ;